+++ Changelog.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-### 0.6-dev
+### 0.6 (2013-04-06)
 
 * When the retrieved mapping from the bower command ```bower list --map``` is invalid, the dependency cache will be deleted
 * The dependency cache warmer no longer creates the cache twice when install_on_warmup is enabled